[Change] separate image routes from index.js
