+++ index.js
@@ -2,26 +2,17 @@ if (process.env.NODE_ENV !== "production") {
     require("dotenv").config();
 }
 
-const { v4: uuidv4 } = require("uuid");
-
 const express = require("express");
 const session = require("express-session");
 const cors = require("cors");
-const fse = require("fs-extra");
-const multer = require("multer");
-const jwt = require("jsonwebtoken");
-const bcrypt = require("bcrypt");
 const path = require("path");
 
 const bodyParser = require("body-parser");
 const cookieParser = require("cookie-parser");
 
-const mailchimp = require("@mailchimp/mailchimp_marketing");
-
-const { dbPoolPromise, checkDB } = require("./mySqlConnection");
-const { verifyJWT, verifyClient } = require("./JWT/jwtMiddleware");
+const { checkDB } = require("./mySqlConnection");
+const { verifyClient } = require("./JWT/jwtMiddleware");
 const { logger } = require("./utils/winstonLogger");
-const { catchRequestError } = require("./utils/catchAsync");
 
 const bookRoutes = require("./routes/bookRoutes");
 const authorRoutes = require("./routes/authorRoutes");
@@ -35,24 +26,9 @@ const ordersRoutes = require("./routes/ordersRoutes");
 const sysRoutes = require("./routes/systemRoutes");
 const messagesRoutes = require("./routes/messagesRoutes");
 const adminRoutes = require("./routes/adminRoutes");
+const imageRoutes = require("./routes/imageRoutes");
 
 const port = process.env.PORT || 3001;
-
-const storage = multer.diskStorage({
-    destination: (req, file, callBack) => {
-        callBack(null, "uploads");
-    },
-    filename: (req, file, callBack) => {
-        callBack(null, `${uuidv4()}-${file.originalname}`);
-    }
-});
-
-const upload = multer({
-    storage: storage,
-    limits: {
-        fileSize: 1024 * 1024 * 5
-    }
-});
 
 const app = express();
 
@@ -113,50 +89,7 @@ app.use("/api/orders", checkDB, verifyClient, ordersRoutes);
 app.use("/api/system", checkDB, sysRoutes);
 app.use("/api/messages", checkDB, verifyClient, messagesRoutes);
 app.use("/api/admin", verifyClient, adminRoutes);
-
-app.post(
-    "/api/images/addimages",
-    checkDB,
-    verifyClient,
-    upload.array("images", 5),
-    (req, res) => {
-        const fileList = req.files;
-        res.send(fileList);
-        fileList.forEach(async (file) => {
-            const [saveImage] = await dbPoolPromise.execute(
-                "INSERT INTO images (id, name, source) VALUES (?, ?, ?)",
-                [file.filename, file.originalname, file.path]
-            );
-        });
-    }
-);
-
-app.post(
-    "/api/images/deleteimages",
-    checkDB,
-    verifyClient,
-    catchRequestError(async (req, res) => {
-        const imgUrl = req.body.url;
-        const newUrl = JSON.stringify(`./${imgUrl.replace(/\\/g, "/")}`);
-        fse.remove(imgUrl, (err) => {
-            if (err) return console.log(err);
-        });
-        const [delImg] = await dbPoolPromise.execute(
-            "DELETE FROM images WHERE source = ?",
-            [imgUrl]
-        );
-    })
-);
-
-app.get(
-    "/api/images/getimages",
-    checkDB,
-    verifyClient,
-    catchRequestError(async (req, res) => {
-        const [imageList] = await dbPoolPromise.execute("SELECT * FROM images");
-        res.send(imageList);
-    })
-);
+app.use("/api/images", checkDB, verifyClient, imageRoutes);
 
 app.get("*", checkDB, (req, res) => {
     res.sendFile(path.resolve(__dirname, "./client", "index.html"));
